+++ src/Core/hco_error_mod.F90
@@ -8,7 +8,7 @@
 ! !DESCRIPTION: Module HCO\_Error\_Mod contains routines and variables
 ! for error handling and logfile messages in HEMCO. It alse contains
 ! definitions of some globally used parameters, such as single/double
-! precision as well as the HEMCO precision which can be either single or
+! precision as well s the HEMCO precision which can be either single or
 ! double. The HEMCO precision is used for almost all HEMCO internal data
 ! arrays and can be changed if required.
 !\\
@@ -183,7 +183,6 @@ CONTAINS
 !
 ! !USES:
 !
-    USE HCO_CharTools_Mod, ONLY : HCO_WordWrapPrint
 #if defined( ESMF_ )
 #include "MAPL_Generic.h"
     USE ESMF
@@ -220,7 +219,7 @@ CONTAINS
 
     ! Specify where to write
     hcoLogLUN = 6
-    IF ( PRESENT(LUN) ) hcoLogLUN = LUN
+    IF ( PRESENT( LUN ) ) hcoLogLUN = LUN
 
     ! Construct error message
 #if defined( ESMF_ )
@@ -266,11 +265,7 @@ CONTAINS
 !
   SUBROUTINE HCO_Warning( ErrMsg, THISLOC, LUN )
 !
-! !USES:
-!
-    USE HCO_CharTools_Mod, ONLY : HCO_WordWrapPrint
-!
-! !INPUT PARAMETERS"
+! !INPUT PARAMETERS:
 !
     CHARACTER(LEN=*), INTENT(IN   )            :: ErrMsg
     CHARACTER(LEN=*), INTENT(IN   ), OPTIONAL  :: THISLOC
@@ -290,11 +285,8 @@ CONTAINS
     !======================================================================
 
     ! Specify where to write
-    IF ( PRESENT(LUN) ) THEN
-       hcoLogLUN = LUN
-    ELSE
-       hcoLogLUN = 6
-    ENDIF
+    hcoLogLUN = 6
+    IF ( PRESENT( LUN ) ) hcoLogLUN = LUN
 
     ! Print warning
     MSG = 'HEMCO WARNING: ' // TRIM( ErrMsg )
@@ -327,10 +319,6 @@ CONTAINS
 ! !INTERFACE:
 !
   SUBROUTINE HCO_MSG( Msg, Sep1, Sep2, LUN )
-!
-! !USES:
-!
-    USE HCO_CharTools_Mod, ONLY : HCO_WordWrapPrint
 !
 ! !INPUT PARAMETERS:
 !
@@ -851,5 +839,182 @@ CONTAINS
     Err%LogIsOpen = .FALSE.
 
   END SUBROUTINE HCO_LogFile_Close
+!EOC
+!------------------------------------------------------------------------------
+!                  GEOS-Chem Global Chemical Transport Model                  !
+!------------------------------------------------------------------------------
+!BOP
+!
+! !IROUTINE: HCO_CountMatches
+!
+! !DESCRIPTION: Counts the number of characters in str1 that match
+!  a character in str2.  This is a dependency of HCO_WordWrapPrint.
+!\\
+!\\
+! !INTERFACE:
+!
+  SUBROUTINE HCO_CountMatches( Str1, Str2, Imat, Locations )
+!
+! !INPUT PARAMETERS:
+!
+    CHARACTER(LEN=*), INTENT(IN) ::  Str1             ! Text to scan
+    CHARACTER(LEN=*), INTENT(IN) ::  Str2             ! Character to match
+!
+! !OUTPUT PARAMETERS:
+!
+    INTEGER,          INTENT(OUT) :: imat             ! Number of matches
+    INTEGER,          OPTIONAL    :: Locations(255)   ! Positions of matches
+!
+! !REVISION HISTORY:
+!     DATE:   JAN. 6, 1995
+!     AUTHOR: R.D. STEWART
+!     COMMENTS: Revised slightly (2-5-1996) so that trailing
+!               blanks in str1 are ignored.  Revised again
+!               on 3-6-1996.
+!  See https://github.com/geoschem/geos-chem for complete history
+!EOP
+!------------------------------------------------------------------------------
+!BOC
+!
+! !LOCAL VARIABLES:
+!
+    ! Scalars
+    INTEGER :: L1, L2, i, j
+    LOGICAL :: again
+
+    ! Arrays
+    INTEGER :: TmpLocations(255)
+
+    ! Initialize
+    TmpLocations = 0
+    L1           = MAX(1,LEN_TRIM(str1))
+    L2           = LEN(str2)
+    imat         = 0
+
+    DO i=1,L1
+       again = .true.
+       j = 1
+       DO WHILE (again)
+          IF (str2(j:j).EQ.str1(i:i)) THEN
+             imat               = imat+1
+             TmpLocations(imat) = i
+             again              = .false.
+          ELSEIF (j.LT.L2) THEN
+             j=j+1
+          ELSE
+             again = .false.
+          ENDIF
+       ENDDO
+    ENDDO
+
+    ! Return positions where matches occured (OPTIONAL)
+    IF ( PRESENT( Locations ) ) Locations = TmpLocations
+
+  END SUBROUTINE HCO_CountMatches
+!EOC
+!------------------------------------------------------------------------------
+!                  GEOS-Chem Global Chemical Transport Model                  !
+!------------------------------------------------------------------------------
+!BOP
+!
+! !IROUTINE: HCO_WordWrapPrint
+!
+! !DESCRIPTION: Prints a text string wrapped to a specified line width.
+!  Useful for displaying error and warning messages.
+!\\
+!\\
+! !INTERFACE:
+!
+  SUBROUTINE HCO_WordWrapPrint( Text, LineWidth, Delimiter, FileLun )
+!
+! !INPUT PARAMETERS:
+!
+    CHARACTER(LEN=*), INTENT(IN) :: Text        ! Text to print
+    INTEGER,          INTENT(IN) :: LineWidth   ! Width (characters) of lines
+    CHARACTER(LEN=1), OPTIONAL   :: Delimiter   ! Delimiter between words
+    INTEGER,          OPTIONAL   :: fileLun     ! LUN of file to write
+!
+! !REMARKS:
+!  (1) This routine must be placed here in hco_error_mod.F90 and not
+!      in hco_chartools_mod.F90 in order to avoid a circular dependency.
+!  (2) The default DELIMITER is the space (" ") character.
+!
+! !REVISION HISTORY:
+!  20 Dec 2015 - R. Yantosca - Initial version
+!  See https://github.com/geoschem/geos-chem for complete history
+!EOP
+!------------------------------------------------------------------------------
+!BOC
+!
+! !LOCAL VARIABLES:
+!
+    ! Scalars
+    INTEGER          :: C, S, B, Matches, Length, LUN
+
+    ! Arrays
+    INTEGER          :: BreakPts(100)
+    INTEGER          :: SpaceLoc(500)
+
+    ! Strings
+    CHARACTER(LEN=1) :: Delim
+
+    !=======================================================================
+    ! WordWrapPrint begins here!
+    !=======================================================================
+
+    ! SpaceLoc is the array of where delimiters (usually the " "
+    ! character) occur in the text, and S is its index.
+    S           = 1
+    SpaceLoc    = 0
+
+    ! BreakPts is the array of where line breaks occur
+    ! and B is its index.
+    BreakPts    = 0
+    B           = 1
+    BreakPts(B) = 1
+
+    ! Delimiter for separating words (will be the space character by default)
+    Delim = ' '
+    IF ( PRESENT( Delimiter ) ) Delim = Delimiter
+
+    ! If writing to file, get the logical unit numbef from the fileLun arg
+    LUN = 6
+    IF ( PRESENT( FileLun ) ) LUN = fileLun
+
+    ! Find the Location of spaces in the text
+    CALL HCO_CountMatches( Text, ' ', Matches, SpaceLoc )
+
+    ! Loop through the number of matches
+    DO
+
+       ! Move to the next delimiter location
+       S = S + 1
+
+       ! Compute the length of the line
+       Length = SpaceLoc(S) - BreakPts(B)
+
+       ! If the length of this segment is greater than the requested
+       ! line length, store the position of this line break
+       IF ( Length > LineWidth ) THEN
+          B           = B             + 1
+          BreakPts(B) = SpaceLoc(S-1) + 1
+       ENDIF
+
+       ! If we have exceeded the number of delimiters in the text, then set
+       ! the last breakpoint at the end of the text and exit the loop.
+       IF ( S > Matches ) THEN
+          B           = B + 1
+          BreakPts(B) = LEN_TRIM( Text ) + 1
+          EXIT
+       ENDIF
+
+    ENDDO
+
+    ! Print each line
+    DO C = 1, B-1
+       WRITE( LUN, '(a)' ) Text( BreakPts(C):BreakPts(C+1)-1 )
+    ENDDO
+
+  END SUBROUTINE HCO_WordWrapPrint
 !EOC
 END MODULE HCO_Error_Mod