+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - TBD
 ### Added
 - Added the functionality of selectively applying gridded scaling factor based on the value of mask ID from a gridded file
-- Added functions `HCO_WordWrapPrint` and `HCO_CountMatches` to `hco_chartools_mod.F90`
+- Added functions `HCO_WordWrapPrint` and `HCO_CountMatches` as `PRIVATE` routines to `hco_error_mod.F90`
 - Added function `IO_ErrMsg` to `hcoio_read_std_mod.F90` to generate a replacement error message for the "cannot find field for current simulation time" message
 
 ### Changed